% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC.fastCUB}
\alias{BIC.fastCUB}
\title{S3 BIC method for class "fastCUB"}
\usage{
\method{BIC}{fastCUB}(object, ...)
}
\arguments{
\item{object}{An object of class "fastCUB"}

\item{...}{Other arguments}
}
\value{
BIC index for the fitted model.
}
\description{
S3 BIC method for objects of class \code{\link{fastCUB}}.
}
\seealso{
\code{\link{logLik}}, \code{\link{fastCUB}}
}
\keyword{package}
