% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.fastCUB}
\alias{coef.fastCUB}
\title{S3 Method: coef for class "fastCUB"}
\usage{
\method{coef}{fastCUB}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{fastCUB}}}

\item{...}{Other arguments}
}
\value{
ML estimates of parameters of the fitted CUB model.
}
\description{
S3 method: coef for objects of class \code{\link{fastCUB}}.
}
\details{
Returns estimated values of coefficients of the fitted model
}
\seealso{
\code{\link{fastCUB}}, \code{\link{summary}}
}
\keyword{package}
