% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.R
\name{decomp}
\alias{decomp}
\title{Louis' identity for the observed information matrix of the incomplete problem for CUB models}
\usage{
decomp(ttau,ordinal,m,param,ai,Y,W)
}
\arguments{
\item{ttau}{Vector of posterior probabilities that each observation has been generated by the first mixture component (feeling)}

\item{ordinal}{Vector of ordinal responses}

\item{m}{Number of ordered categories}

\item{param}{Vector of estimable parameters}

\item{ai}{Auxiliary scalars}

\item{Y}{Matrix of selected covariates for explaining the uncertainty component}

\item{W}{Matrix of selected covariates for explaining the feeling component}
}
\value{
A list of the matrices in Louis' identity for the observed information matrix
}
\description{
Compute the matrices involved in Louis' identity for the observed information matrix
}
\seealso{
\code{\link{fastCUB}}
}
\keyword{internal}
