% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effe01.R
\name{effe01}
\alias{effe01}
\title{Auxiliary function for the log-likelihood estimation of CUB models}
\usage{
effe01(gama, esterno01, m)
}
\arguments{
\item{gama}{Vector of the starting values of the parameters to be estimated}

\item{esterno01}{A matrix binding together the vector of the posterior probabilities
that each observation has been generated by the first component distribution of the mixture, 
the ordinal data and the matrix of the selected covariates accounting for an intercept term}
}
\description{
Compute the opposite of the scalar function that is maximized when running 
the E-M algorithm for CUB models with covariates for the feeling parameter.
}
\details{
It is called as an argument for optim within CUB function for models with covariates for
feeling or for both feeling and uncertainty
}
\keyword{internal}
