% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effe10.R
\name{effe10}
\alias{effe10}
\title{Auxiliary function for the log-likelihood estimation of CUB models}
\usage{
effe10(bet, esterno10)
}
\arguments{
\item{bet}{Vector of the starting values for the parameters to be estimated}

\item{esterno10}{A matrix binding together the matrix of the selected covariates
(accounting for an intercept term) and a vector (whose length equals the number of observations)
of the posterior probabilities that each observation has been generated by the first component
distribution of the mixture}
}
\description{
Compute the opposite of the scalar function that is maximized when running
the E-M algorithm for CUB models with covariates for the uncertainty parameter.
}
\details{
It is called as an argument for optim within CUB function for models with covariates for
uncertainty or for both feeling and uncertainty
}
\keyword{internal}
