% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcub00.R
\name{fastcub00}
\alias{fastcub00}
\title{Main function for CUB models without covariates}
\usage{
fastcub00(m,ordinal,starting=NULL,maxiter,toler,iterc=3,invgen=TRUE,verbose=FALSE)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{starting}{Starting values for the algorithm}

\item{maxiter}{Maximum number of iterations allowed for running the optimization algorithm}

\item{toler}{Fixed error tolerance for final estimates}

\item{iterc}{Iteration from which the acceleration strategy starts}

\item{invgen}{Logical: should the recursive formula for the inverse of the information matrix be considered? (Default is TRUE)}

\item{verbose}{Logical: should messages about acceleration steps be printed out? (Default is FALSE)}
}
\value{
An object of the class "fastCUB"
}
\description{
Function to estimate and validate a CUB model without covariates for given ordinal responses.
}
\seealso{
\code{\link{probbit}}, \code{\link{probcub00}}
}
\keyword{internal}
