% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.fastCUB}
\alias{fitted.fastCUB}
\title{S3 method "fitted" for class "fastCUB"}
\usage{
\method{fitted}{fastCUB}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{fastCUB}}}

\item{...}{Other arguments}
}
\description{
S3 method fitted for objects of class \code{\link{fastCUB}}.
}
\details{
Returns the fitted probability distribution for GEM models with no covariates. If only one dichotomous 
covariate is included in the model to explain some components, it returns the fitted probability distribution for each profile.
}
\seealso{
\code{fastCUB}
}
\keyword{package}
