% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hadprod.R
\name{Hadprod}
\alias{Hadprod}
\title{Hadamard product of a matrix with a vector}
\usage{
Hadprod(Amat, xvett)
}
\arguments{
\item{Amat}{A generic matrix}

\item{xvett}{A generic vector}
}
\value{
A matrix of the same dimensions as Amat
}
\description{
Return the Hadamard product between the given matrix and vector: this operation corresponds
to multiply every row of the matrix by the corresponding element of the vector, and it is equivalent to the
standard matrix multiplication to the right with the diagonal matrix whose diagonal is the given vector.
It is possible only if the length of the vector equals the number of rows of the matrix, otherwise it prints
an error message.
}
\details{
It is an auxiliary function needed for computing the variance-covariance matrix of the estimated model
with covariates.
}
\keyword{internal}
