% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invmatgen.R
\name{invmatgen}
\alias{invmatgen}
\title{Recursive computation of the inverse of a matrix}
\usage{
invmatgen(G,H,listE)
}
\arguments{
\item{G}{Primary matrix for the sum decomposition of $G+H$}

\item{H}{Secondary matrix for the sum decomposition of $G+H$}

\item{listE}{Auxiliary matrices that sum up to H}
}
\value{
The inverse of matrix G + H computed recursively thanks to matrices listed in listE
}
\description{
Compute the variance-covariance matrix of the incomplete score vector involved in Louis' identity for the observed information matrix
}
\references{
Miller K. (1981).  On the inverse of the sum of matrices, 
\emph{Mathematics Magazine}, \bold{54}, 67--72 \cr
}
\seealso{
\code{\link{fastCUB}}
}
\keyword{stats}
