% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.fastCUB}
\alias{logLik.fastCUB}
\title{logLik S3 Method for class "fastCUB"}
\usage{
\method{logLik}{fastCUB}(object, ...)
}
\arguments{
\item{object}{An object of class "fastCUB"}

\item{...}{Other arguments}
}
\value{
Log-likelihood at the final ML estimates for parameters of the fitted fastCUB model.
}
\description{
S3 method: logLik() for objects of class "fastCUB".
}
\seealso{
\code{\link{fastCUB}}
}
\keyword{package}
