% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikcub00.R
\name{loglikcub00}
\alias{loglikcub00}
\title{Log-likelihood function of a CUB model without covariates}
\usage{
loglikcub00(m, freq, pai, csi)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{freq}{Vector of the absolute frequency distribution}

\item{pai}{Uncertainty parameter}

\item{csi}{Feeling parameter}
}
\description{
Compute the log-likelihood function of a CUB model without covariates for a given
absolute frequency distribution.
}
\keyword{internal}
