% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikcub0q.R
\name{loglikcub0q}
\alias{loglikcub0q}
\title{Log-likelihood function of a CUB model with covariates for the feeling component}
\usage{
loglikcub0q(m, ordinal, W, pai, gama)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{W}{Matrix of selected covariates for explaining the feeling component}

\item{pai}{Uncertainty parameter}

\item{gama}{Vector of parameters for the feeling component, with length NCOL(W) + 1 to account for
an intercept term (first entry of gama)}
}
\description{
Compute the log-likelihood function of a CUB model fitting ordinal data, with \eqn{q}
covariates for explaining the feeling component.
}
\keyword{internal}
