% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikcubp0.R
\name{loglikcubp0}
\alias{loglikcubp0}
\title{Log-likelihood function of a CUB model with covariates for the uncertainty component}
\usage{
loglikcubp0(m, ordinal, Y, bbet, ccsi)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{ordinal}{Vector of ordinal responses}

\item{Y}{Matrix of selected covariates for explaining the uncertainty component}

\item{bbet}{Vector of parameters for the uncertainty component, with length equal to
NCOL(Y)+1 to account for an intercept term (first entry of bbet)}

\item{ccsi}{Feeling parameter}
}
\description{
Compute the log-likelihood function of a CUB model fitting ordinal responses with covariates
for explaining the uncertainty component.
}
\keyword{internal}
