% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parnames.R
\name{parnames}
\alias{parnames}
\title{Generic function for coefficient names}
\usage{
parnames(object)
}
\arguments{
\item{object}{An object of class "fastCUB"}
}
\value{
Parameter names
}
\description{
Generic function for names of parameter estimates of object of class "fastCUB".
}
\seealso{
\code{\link{summary}}
}
\keyword{internal}
