% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fastCUB}
\alias{print.fastCUB}
\title{S3 method: print for class "fastCUB"}
\usage{
\method{print}{fastCUB}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{fastCUB}}}

\item{...}{Other arguments}
}
\value{
Brief summary results of the fitting procedure, including parameter estimates, their standard errors and
the executed call.
}
\description{
S3 method print for objects of class \code{\link{fastCUB}}.
}
\keyword{package}
