% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probbit.R
\name{probbit}
\alias{probbit}
\title{Probability distribution of a shifted Binomial random variable}
\usage{
probbit(m,csi)
}
\arguments{
\item{m}{Number of ordinal categories}

\item{csi}{Feeling parameter}
}
\value{
The vector of the probability distribution of a shifted Binomial model.
}
\description{
Return the shifted Binomial probability distribution.
}
\examples{
m<-7
csi<-0.7
pr<-probbit(m,csi)
plot(1:m,pr,type="h",main="Shifted Binomial probability distribution",xlab="Categories")
points(1:m,pr,pch=19)
}
\seealso{
\code{\link{bitcsi}}, \code{\link{probcub00}}
}
\keyword{distribution}
