% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.fastCUB}
\alias{vcov.fastCUB}
\title{S3 method  vcov() for class "fastCUB"}
\usage{
\method{vcov}{fastCUB}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{fastCUB}}}

\item{...}{Other arguments}
}
\value{
Variance-covariance matrix of the final ML estimates for parameters of the fitted CUB model, 
according to the logit transform also when covariates are not included for uncertainty or feeling parameters. 
It is computed on the basis of Louis' identity within the EM algorithm.
}
\description{
S3 method: vcov for objects of class \code{\link{fastCUB}}.
}
\seealso{
\code{\link{fastCUB}}
}
\keyword{package}
